requirements.txt
Flask==3.0.2
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
azure-cosmos==4.5.1
azure-identity==1.16.1
azure-search-documents==11.6.0b3
azure-monitor-opentelemetry==1.6.0